Still writing IR

+++ x86BackEnd/BackEnd/IR/TypeDefs.h
@@ -1,4 +1,4 @@
 #pragma once 
 
 typedef const char* name_t;
-typedef int         data_t;
+typedef double         data_t;
